add changes requested by Ferrão
