+++ src/main/java/com/codeforall/simplegraphics/graphics/Colorable.java
@@ -1,0 +1,15 @@
+package com.codeforall.simplegraphics.graphics;
+
+/**
+ * Methods for modifying a shape color
+ * The color of the shape is used when drawing or filling
+ */
+public interface Colorable {
+
+    /**
+     * Sets the color of the shape
+     *
+     * @param color the new color
+     */
+    void setColor(Color color);
+}
