+++ src/main/java/com/codeforall/simplegraphics/graphics/Fillable.java
@@ -1,0 +1,12 @@
+package com.codeforall.simplegraphics.graphics;
+
+/**
+ * Methods for filling a shape
+ */
+public interface Fillable {
+
+    /**
+     * Paints the shape with the current shape color
+     */
+    void fill();
+}
