+++ src/main/java/com/codeforall/simplegraphics/keyboard/KeyboardEventType.java
@@ -1,0 +1,11 @@
+package com.codeforall.simplegraphics.keyboard;
+
+/**
+ * The type of events supported by the Keyboard
+ * @see Keyboard
+ */
+public enum KeyboardEventType {
+
+    KEY_PRESSED,
+    KEY_RELEASED
+}
