+++ src/main/java/com/codeforall/simplegraphics/mouse/MouseEventType.java
@@ -1,0 +1,10 @@
+package com.codeforall.simplegraphics.mouse;
+
+public enum MouseEventType {
+
+    MOUSE_CLICKED,
+    MOUSE_MOVED;
+
+    private MouseEventType() {
+    }
+}
