--- src/main/java/org/academiadecodigo/simplegraphics/graphics/Fillable.java
@@ -1,12 +1,0 @@
-package org.academiadecodigo.simplegraphics.graphics;
-
-/**
- * Methods for filling a shape
- */
-public interface Fillable {
-
-    /**
-     * Paints the shape with the current shape color
-     */
-    void fill();
-}
