--- src/main/java/org/academiadecodigo/simplegraphics/keyboard/KeyboardEventType.java
@@ -1,11 +1,0 @@
-package org.academiadecodigo.simplegraphics.keyboard;
-
-/**
- * The type of events supported by the Keyboard
- * @see Keyboard
- */
-public enum KeyboardEventType {
-
-    KEY_PRESSED,
-    KEY_RELEASED
-}
