--- src/main/java/org/academiadecodigo/simplegraphics/mouse/MouseEventType.java
@@ -1,10 +1,0 @@
-package org.academiadecodigo.simplegraphics.mouse;
-
-public enum MouseEventType {
-
-    MOUSE_CLICKED,
-    MOUSE_MOVED;
-
-    private MouseEventType() {
-    }
-}
